+++ apps/web/app/shortform/page.tsx
@@ -1,1 +1,3 @@
+export const dynamic = 'force-dynamic';
+
 export { default } from '@/screens/shortform/ui/ShortformPage';
